converted from service back to util
