+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/segmenting/VideoMediaSegmenter.java
@@ -47,7 +47,7 @@ import org.mitre.mpf.wfm.data.entities.transients.Track;
 import org.mitre.mpf.wfm.enums.MpfConstants;
 import org.mitre.mpf.wfm.util.AggregateJobPropertiesUtil;
 import org.mitre.mpf.wfm.util.MediaRange;
-import org.mitre.mpf.wfm.service.TopQualitySelectionService;
+import org.mitre.mpf.wfm.util.TopQualityUtil;
 import org.mitre.mpf.wfm.data.InProgressBatchJobsService;
 import org.mitre.mpf.wfm.util.UserSpecifiedRangesUtil;
 import org.slf4j.Logger;
@@ -168,7 +168,7 @@ public class VideoMediaSegmenter implements MediaSegmenter {
             stopFrame = track.getEndOffsetFrameInclusive();
         }
         else {
-            includedDetections = TopQualitySelectionService.getTopQualityDetections(
+            includedDetections = TopQualityUtil.getTopQualityDetections(
                     track.getDetections(), topQualityCount, topQualitySelectionProp);
             var frameSummaryStats = includedDetections.stream()
                     .mapToInt(Detection::getMediaOffsetFrame)
