--- trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/service/TopQualitySelectionService.java
@@ -1,121 +1,0 @@
-/******************************************************************************
- * NOTICE                                                                     *
- *                                                                            *
- * This software (or technical data) was produced for the U.S. Government     *
- * under contract, and is subject to the Rights in Data-General Clause        *
- * 52.227-14, Alt. IV (DEC 2007).                                             *
- *                                                                            *
- * Copyright 2023 The MITRE Corporation. All Rights Reserved.                 *
- ******************************************************************************/
-/******************************************************************************
- * Copyright 2023 The MITRE Corporation                                       *
- *                                                                            *
- * Licensed under the Apache License, Version 2.0 (the "License");            *
- * you may not use this file except in compliance with the License.           *
- * You may obtain a copy of the License at                                    *
- *                                                                            *
- *    http://www.apache.org/licenses/LICENSE-2.0                              *
- *                                                                            *
- * Unless required by applicable law or agreed to in writing, software        *
- * distributed under the License is distributed on an "AS IS" BASIS,          *
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   *
- * See the License for the specific language governing permissions and        *
- * limitations under the License.                                             *
- ******************************************************************************/
-
-
-package org.mitre.mpf.wfm.service;
-
-import static java.util.stream.Collectors.toMap;
-
-import java.lang.NumberFormatException;
-import java.lang.NullPointerException;
-import java.util.Collection;
-import java.util.Comparator;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.PriorityQueue;
-import java.util.function.ToDoubleFunction;
-import java.util.stream.Stream;
-
-import javax.inject.Inject;
-
-import org.mitre.mpf.wfm.data.InProgressBatchJobsService;
-import org.mitre.mpf.wfm.data.entities.persistent.BatchJob;
-import org.mitre.mpf.wfm.data.entities.persistent.Media;
-import org.mitre.mpf.wfm.data.entities.persistent.SystemPropertiesSnapshot;
-import org.mitre.mpf.wfm.data.entities.transients.Detection;
-import org.mitre.mpf.wfm.enums.BatchJobStatusType;
-import org.mitre.mpf.wfm.enums.IssueCodes;
-import org.mitre.mpf.wfm.enums.IssueSources;
-import org.mitre.mpf.wfm.enums.MpfConstants;
-import org.mitre.mpf.wfm.enums.MpfHeaders;
-import org.mitre.mpf.rest.api.pipelines.Action;
-import org.mitre.mpf.wfm.util.AggregateJobPropertiesUtil;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Component;
-
-import com.google.common.collect.Streams;
-
-public class TopQualitySelectionService {
-
-    private TopQualitySelectionService() {}
-
-    public static Detection getTopQualityItem(
-            Collection<Detection> detections, String qualityProperty) {
-        return detections.stream()
-            .max(getMaxQualityComparator(d -> getQuality(d, qualityProperty.toLowerCase())))
-            .orElse(null);
-    }
-
-    public static Collection<Detection> getTopQualityDetections(
-            Collection<Detection> allDetections, int topQualityCount,
-            String qualityProperty) {
-
-        if (topQualityCount <= 0 || topQualityCount >= allDetections.size()) {
-            return allDetections;
-        }
-        var qualityComparator = getMaxQualityComparator(((Detection d) -> getQuality(d, qualityProperty)));
-
-        var topDetections = new PriorityQueue<>(topQualityCount, qualityComparator);
-
-        var allDetectionsIter = allDetections.iterator();
-        for (int i = 0; i < topQualityCount; i++) {
-            topDetections.add(allDetectionsIter.next());
-        }
-
-        while (allDetectionsIter.hasNext()) {
-            Detection detection = allDetectionsIter.next();
-            // Check if current detection is less than the minimum top detection so far
-            if (qualityComparator.compare(detection, topDetections.peek()) > 0) {
-                topDetections.poll();
-                topDetections.add(detection);
-            }
-        }
-        return topDetections;
-    }
-
-    public static double getQuality(Detection det, String qualityProperty) {
-        try {
-            if (qualityProperty.toLowerCase().equals("confidence"))
-                return det.getConfidence();
-            else
-                return Double.parseDouble(det.getDetectionProperties().get(qualityProperty));
-        }
-        catch(NumberFormatException e) {
-            String errMsg = "The quality selection property \"" + qualityProperty + "\" could not be converted to a double value: " + det.getDetectionProperties().get(prop);
-            throw new NumberFormatException(errMsg);
-        }
-        catch(NullPointerException e) {
-            String errMsg = "The value of quality selection property \"" + qualityProperty + "\" is null.";
-            throw new NumberFormatException(errMsg);
-       }
-    }
-
-    private static <T extends Comparable<T>>
-            Comparator<T> getMaxQualityComparator(ToDoubleFunction<T> qualityGetter) {
-        return Comparator.comparingDouble(qualityGetter)
-                .thenComparing(Comparator.reverseOrder());
-    }
-}
